Home

AdaMeanShift.jl

Julia package to perform multidimensional and scale-adaptive meanshift over density tensors (Images, Histograms, ...)

AdaMeanShift.meanshift!

AdaMeanShift.meanshift! — Method.

meanshift!(M, P, h, w, hmax; isotropy=½, maxit=∞, rtol=√ϵ, smoothing=1)

Performs MeanShift on a swarm of particles over a given density matrix locating all modes and their scale.

Arguments

`M` is the density tensor over which particles evolve.
`P` is a julia vector of StaticVectors describing all positions.
`h` is a julia vector of StaticVectors describing all standard deviations.
`w` is a julia vector which will contain the new modes intensity estimates.
`hmax` is the maximum norm that standard deviations "h" can have.

Keyword Arguments

`isotropy` is a scalar [0,1] set to 1 for isotropic kernel,
                   < 1 anisotropy along coordinate axis is allowed.
`maxit` is a integer for the maximum number of meanshift iterations.
`rtol` is the absolute tolerance to declare a particle as converged.
`smoothing` is a regularization term for density tensors with noise.

source