API

AdaMeanShift.jl

Julia package to perform multidimensional and scale-adaptive meanshift over density tensors (Images, Histograms, ...)

AdaMeanShift.atomic_intensity

AdaMeanShift.atomic_meanshift

AdaMeanShift.meanshift!

AdaMeanShift.atomic_intensity — Method.

    atomic_intensity(M,p,h) -> NamedTuple(intensity,numpoints,ndims)

Calculates the integrated intensity on the density tensor M over a ellipsoidal region caracterized by p region center, h semiaxis

Arguments

M density tensor (e.g. matrix N*N)

p region center (e.g. 2D vector)

h semiaxis lengths (e.g. 2D vector)

Note that both p,h should be of type Vector or StaticVector with the same length.

source

AdaMeanShift.atomic_meanshift — Method.

    atomic_meanshift(M,p,h,isotropy,smoothing) -> NamedTuple(pnew,hnew,modeval,delta)

Performs a single iteration of meanshift over a single particle

Arguments

M density tensor

p initial position for test particle

h initial particle standard deviation

isotropy scalar 0 -> 1, 0 is completely anisotropic, completely isotropic

smoothing scalar 0 -> 1, regularization term for noisy density tensors

source

AdaMeanShift.meanshift! — Method.

    meanshift!(M, P, h, w, hmax; isotropy=½, maxit=∞, rtol=√ϵ, smoothing=1)
    meanshift_nonadaptive!(M, P, h, w; maxit=∞, rtol=√ϵ, smoothing=1)

Performs MeanShift on a swarm of particles over a given density matrix locating all modes and their scale. The non adaptive method disables updating of the h vectors, and thus depends on fewer parameters.

Arguments

M is the density tensor over which particles evolve.

P is a julia vector of StaticVectors describing all positions.

h is a julia vector of StaticVectors describing all standard deviations.

w is a julia vector which will contain the new modes intensity estimates.

hmax is a real number for the maximum norm allowed to the h vectors.

Keyword Arguments

isotropy is a scalar [0,1] set to 1 for isotropic kernel, < 1 anisotropy along coordinate axis is allowed.

maxit is a integer for the maximum number of meanshift iterations.

rtol is the absolute tolerance to declare a particle as converged.

smoothing is a regularization term for density tensors with noise.

source