Home

AdaMeanShift.jl

AdaMeanShift.runms!

AdaMeanShift.runms! — Method.

runms!(M,P,h,w,hmax,isotropy,maxit,rtol) –– Performs MeanShift on a swarm of particles over a given density matrix locating all modes and their scale.

* M is the density matrix over which particles evolve
* P is a julia vector of StaticVectors describing all positions
* h is a julia vector of StaticVectors describing all standard deviations
* w is a julia vector which will contain the new modes intensity estimates
* hmax is the maximum norm that standard deviations "h" can have
* isotropy=0.5 is a scalar [0,1] set to 1 for isotropic kernel,
                   < 1 anisotropy along coordinate axis is allowed
* maxit=Inf is a integer for the maximum number of meanshift iterations
* rtol=√ϵ is the absolute tolerance to declare a particle as converged

source